+++ src/components/CardPost.tsx
@@ -3,8 +3,9 @@
 import { formatedDate } from '@/utils/dateUtils'
 import React, { HTMLAttributes } from 'react'
 import CardMenu from './CardMenu'
-import { Operations, Post } from '@/types'
+import { Operations } from '@/types'
 import { useDialogContext } from '@/context/dialogContext'
+import { Post } from '@/types/Post'
 
 interface CardPostProps extends HTMLAttributes<HTMLDivElement> {
   post: Post
@@ -19,11 +20,11 @@ const CardPost = ({ post }: CardPostProps) => {
       className='relative flex flex-col px-4 py-2 w-[100%] max-w-[500px] border rounded-md cursor-pointer transition-all duration-300 hover:shadow-md'
     >
 
-      <CardMenu id={id} />
+      <CardMenu id={id!} />
 
       <div
         className='h-[125px]'
-        onClick={() => dispatch({ type: Operations.READ, payload: id })}
+        onClick={() => dispatch({ type: Operations.READ, payload: id! })}
       >
 
         <h1 className='text-lg font-semibold'>
@@ -35,7 +36,7 @@ const CardPost = ({ post }: CardPostProps) => {
         </p>
 
         <p className='absolute bottom-2 right-4'>
-          {formatedDate(createdAt)}
+          {formatedDate(createdAt!)}
         </p>
       </div>
     </div>
