+++ src/utils/dateUtils.ts
@@ -1,4 +1,4 @@
-import { Post } from "@/types"
+import { Post } from "@/types/Post"
 
 export const formatedDate = (date: string): string => {
   return new Date(date).toLocaleString('pt-BR', {
@@ -12,5 +12,5 @@ export const formatedDateWTime = (date: string | undefined): string | undefined 
 
 
 export const sortDate = (posts: Array<Post>): Array<Post> => {
-  return posts.sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
+  return posts.sort((a, b) => new Date(b.createdAt!).getTime() - new Date(a.createdAt!).getTime())
 }
